refactor(api): split conversations.ts into focused route modules
Replace monolithic conversations.ts (650 lines, 11 endpoints) with a
directory of focused modules:
- crud.ts: POST /, GET /, GET /:id, PUT /:id, DELETE /:id
- messages.ts: POST /:id/messages, GET /:id/messages
- search.ts: GET /search, GET /by-external-id/:externalId
- bulk.ts: POST /bulk-delete, POST /export
- index.ts: combines sub-routers, applies auth + rate-limit once

Also adopts existing loadConversation/notFound helpers from lib/helpers
to eliminate repeated conversation lookup + 404 boilerplate across 5
endpoints.
